+++ CHANGES.rst
@@ -9,6 +9,7 @@ v3.1.0 (xxxx-xx-xx)
 
 - Add ``fill``.
 - Add ``in_range``.
+- Add ``start_case``.
 
 
 .. _changelog-v3.0.0: